+++ src/configs/recommended.js
@@ -11,11 +11,11 @@ module.exports = {
   extends: ['plugin:github/react'],
   rules: {
     'primer-react/direct-slot-children': 'error',
-    'primer-react/no-deprecated-colors': 'warn',
     'primer-react/no-system-props': 'warn',
     'primer-react/a11y-tooltip-interactive-trigger': 'error',
     'primer-react/new-color-css-vars': 'error',
     'primer-react/a11y-explicit-heading': 'error',
+    'primer-react/no-deprecated-props': 'warn',
     'primer-react/new-color-css-vars-have-fallback': 'error',
     'primer-react/use-next-tooltip': 'error',
   },