+++ src/rules/__tests__/no-system-props.test.js
@@ -22,6 +22,12 @@ ruleTester.run('no-system-props', rule, {
     `import {Button} from '@primer/react'; <Button variant="large" />`,
     `import {Button} from '@primer/react'; <Button size="large" />`,
     `import {ActionMenu} from '@primer/react'; <ActionMenu.Overlay width="large" />`,
+    {code: `<img width="200px" />`, options: [{skipImportCheck: true}]},
+    {code: `<Placeholder width="200px" />`, options: [{skipImportCheck: true, ignoreNames: ['Placeholder']}]},
+    {
+      code: `<Placeholder.Header width="200px" />`,
+      options: [{skipImportCheck: true, ignoreNames: ['Placeholder.Header']}],
+    },
   ],
   invalid: [
     {